Impose a(1,1) <= a(2,2) <= a(3,3) <= a(4,4)
It is safe to assume that the elements along the diagonal are in order
since the problem is invariant under similarity transformation with
permutation matrices.

This reduces the computing time by a factor of 14 (exact value: 10000/715).
The results remain unchanged.

```matlab
A_max =

     2    10    10     7
    10     9     1     9
    10     1    10    10
     7     9    10    10

cond_max =

   7.6119e+04

A_max_pd =

     9     1     1     5
     1    10     1     9
     1     1    10     1
     5     9     1    10

cond_max_pd =

   3.5529e+04
```